지점 계정 승인 및 매장 운영 관리 기능 구현

+++ backend/src/main/java/com/kiosk/domain/branch/Branch.java
@@ -63,7 +63,7 @@ public class Branch {
     private Boolean isBusy = false;
 
     @Column(name = "estimated_wait_minutes")
-    private Integer estimatedWaitMinutes;
+    private Byte estimatedWaitMinutes;
 
     @Column(name = "kiosk_code", length = 50)
     private String kioskCode;
